scripts/requirements.txt
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
mysql-connector-python
mysqlclient
apache-airflow-providers-mysql
